refactor: 배너이미지 utils->constants로 폴더변경
