+++ frontend/src/constants/banners.ts
