+++ frontend/src/pages/MainPage/MainPage.tsx
@@ -9,8 +9,8 @@ import Footer from '@/components/common/Footer/Footer';
 import Header from '@/components/common/Header/Header';
 import MainMobileHeader from '@/pages/MainPage/components/MobileHeader/MobileHeader';
 import Banner from '@/pages/MainPage/components/Banner/Banner';
-import { DesktopBannerImageList } from '@/utils/banners';
-import { MobileBannerImageList } from '@/utils/banners';
+import { DesktopBannerImageList } from '@/constants/banners';
+import { MobileBannerImageList } from '@/constants/banners';
 import { Club } from '@/types/club';
 import * as Styled from './MainPage.styles';
